level 2-9 // modify 생성자

+++ src/AddOperator.java
@@ -1,5 +1,5 @@
-public class AddOperator {
-    public int operate(int x, int y) {
+public class AddOperator  {
+    public double operate(int x, int y) {
         return x + y;
     }
 }
