+++ src/MultiOperator.java
@@ -1,5 +1,5 @@
 public class MultiOperator {
-    public int operate (int x, int y){
+    public double operate (int x, int y){
         return x * y;
     }
 }
